removal of transaction register and fix notificaiton width
